+++ src/init.c
@@ -10,6 +10,7 @@ FIXME: Check these declarations against the C/Fortran source code.
 */
 
 /* .Fortran calls */
+extern void F77_NAME(countf)(void *, void *, void *, void *, void *);
 extern void F77_NAME(covwf)(void *, void *, void *, void *, void *, void *, void *, void *);
 extern void F77_NAME(crossprodf)(void *, void *, void *, void *, void *, void *);
 extern void F77_NAME(d2norm)(void *, void *, void *, void *);
@@ -107,6 +108,7 @@ extern void F77_NAME(softmax)(void *, void *, void *, void *, void *, void *);
 extern void F77_NAME(uncholf)(void *, void *, void *, void *, void *);
 
 static const R_FortranMethodDef FortranEntries[] = {
+    {"countf",     (DL_FUNC) &F77_NAME(countf),      5},
     {"covwf",      (DL_FUNC) &F77_NAME(covwf),       8},
     {"crossprodf", (DL_FUNC) &F77_NAME(crossprodf),  6},
     {"d2norm",     (DL_FUNC) &F77_NAME(d2norm),      4},